ai-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pgvector==0.2.5
sentence-transformers==2.5.1
scikit-learn==1.4.1.post1
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.6.3
pydantic-settings==2.2.1
